requirements.txt
discord.py
jishaku
datetime
typing
requests
asyncio
pymongo
better-ipc
roblox
python-dotenv
rollbar
PyYaml